refactor: return an error on unsupported verb instead of panic
