+++ builq.go
@@ -11,6 +11,7 @@ type Builder struct {
 	query strings.Builder
 	args  []any
 	pp    placeholderProvider
+	err   error // the first error occurred while building the query.
 }
 
 func NewIterBuilder(placeholder string) Builder {
@@ -26,6 +27,10 @@ func NewStaticBuilder(placeholder string) Builder {
 }
 
 func (b *Builder) Appendf(format string, args ...any) *Builder {
+	if b.err != nil {
+		return b // return earlier if there is already an error.
+	}
+
 	if b.pp == nil {
 		b.pp = &iterProvider{p: "$"}
 	}
@@ -38,7 +43,12 @@ func (b *Builder) Appendf(format string, args ...any) *Builder {
 	fmt.Fprintf(&b.query, format+"\n", wargs...)
 
 	for _, warg := range wargs {
-		if arg := warg.(*argument); arg.forQuery {
+		arg := warg.(*argument)
+		if err := arg.err; err != nil {
+			b.err = err
+			break
+		}
+		if arg.forQuery {
 			b.args = append(b.args, arg.value)
 		}
 	}
@@ -47,14 +57,18 @@ func (b *Builder) Appendf(format string, args ...any) *Builder {
 }
 
 func (b *Builder) Build() (string, []any, error) {
+	if err := b.err; err != nil {
+		return "", nil, err
+	}
 	return b.query.String(), b.args, nil
 }
 
 // argument is a wrapper for Printf-style arguments that implements fmt.Formatter.
 type argument struct {
 	value    any
-	forQuery bool // is it a query argument?
-	pp       placeholderProvider
+	forQuery bool                // is it a query argument?
+	pp       placeholderProvider // the source of the next placeholder.
+	err      error               // an error occurred during the Format call.
 }
 
 // Format implements the fmt.Formatter interface.
@@ -68,7 +82,8 @@ func (a *argument) Format(s fmt.State, v rune) {
 		a.forQuery = true
 		fmt.Fprint(s, a.pp.Next())
 	default:
-		panic(fmt.Sprintf("unsupported verb %c", v))
+		a.err = fmt.Errorf("builq: unsupported verb %c", v)
+		// panic(a.err) // this panic will be caught and written to s by the fmt code.
 	}
 }
